backend/requirements.txt
fastapi
uvicorn
langgraph
pydantic
python-dotenv
openai
pymongo
requests
finnhub-python
websockets>=13.0
urllib3>=2.2.2
firebase-admin>=6.0.0
pyotp>=2.8.0
fastapi-security>=0.2.0
qrcode>=7.0
fpdf2